refactor: cool-links with an S
