+++ src/modules/coolLinksManagement/summarizeCoolPages.ts
