+++ src/modules/modules.ts
@@ -1,3 +1,4 @@
+import { coolLinksManagement } from './coolLinksManagement/coolLinksManagement.module';
 import { fart } from './fart/fart.module';
 import { patternReplace } from './patternReplace/patternReplace.module';
 import { voiceOnDemand } from './voiceOnDemand/voiceOnDemand.module';
@@ -5,5 +6,6 @@ import { voiceOnDemand } from './voiceOnDemand/voiceOnDemand.module';
 export const modules = {
   fart,
   voiceOnDemand,
+  coolLinksManagement,
   patternReplace,
 };
